As I was saying, added a bunch of models + seeders. Also added in ApiTokens and better unit tests.

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -14,9 +14,12 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->increments('id');
-            $table->string('name');
+            $table->string('firstname');
+            $table->string('lastname');
+            $table->integer('image_id')->unsigned()->nullable();
             $table->string('email')->unique();
-            $table->string('password');
+            $table->string('password', 60);
+            $table->timestamp('last_login_at')->nullable();
             $table->rememberToken();
             $table->timestamps();
         });
@@ -29,6 +32,9 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::drop('users');
+        if (Schema::hasTable('users')) 
+        {
+            Schema::drop('users');
+        }
     }
 }
